Removed "sudo" from docker commands
sudo was used only for running docker commands.

We should avoid to ask for sudo access if it
is not needed. In this case non-root user can
call docker commands without asking for sudo.
It just needs to be properly set after docker
installation. More about that in README.md

Signed-off-by: Martin Sunal <msunal@frinx.io>